fix: 단어장 추가 및 업데이트 시 중복 체크
- 중복 시 메세지와 함께 409 반환

+++ mobidic_spring/src/main/java/com/kimtaeyang/mobidic/repository/VocabRepository.java
@@ -15,5 +15,5 @@ public interface VocabRepository extends JpaRepository<Vocab, UUID> {
 
     Optional<Vocab> findByTitle(String title);
 
-    long countByTitleAndIdNot(String title, UUID id);
+    int countByTitleAndIdNot(String title, UUID id);
 }
